Make Telemetry an interface
In `monaco-yaml` I’m interested in stubbing the Telemetry option.
However, currently everywhere this is used uses a type annotation based
on the class. This means the implementation must match this class,
including private properties. By accepting an interface instead, it’s
possible to pass alternative implementations.
