+++ test/utils/testsTypes.ts
@@ -16,7 +16,8 @@ import {
   DeleteFilesParams,
 } from 'vscode-languageserver-protocol';
 import { Connection, RemoteWorkspace } from 'vscode-languageserver';
-import { Telemetry, TelemetryEvent } from '../../src/languageserver/telemetry';
+import { Telemetry } from '../../src/languageserver/telemetry';
+import { TelemetryEvent } from '../../src/languageservice/telemetry';
 
 /* eslint-disable @typescript-eslint/no-explicit-any */
 /* eslint-disable @typescript-eslint/no-unused-vars */
